Fix Tuple.columnName contract.
