+++ modules/client-handler/src/main/java/org/apache/ignite/client/handler/requests/table/ClientHandlerTuple.java
@@ -25,6 +25,7 @@ import org.apache.ignite.internal.schema.Column;
 import org.apache.ignite.internal.schema.SchemaAware;
 import org.apache.ignite.internal.schema.SchemaDescriptor;
 import org.apache.ignite.internal.type.NativeTypeSpec;
+import org.apache.ignite.lang.util.IgniteNameUtils;
 import org.apache.ignite.sql.ColumnType;
 import org.apache.ignite.table.Tuple;
 import org.jetbrains.annotations.Nullable;
@@ -76,7 +77,7 @@ class ClientHandlerTuple extends MutableTupleBinaryTupleAdapter implements Schem
     /** {@inheritDoc} */
     @Override
     protected int binaryTupleIndex(String columnName) {
-        Column column = schema.column(columnName);
+        Column column = schema.column(IgniteNameUtils.parseSimpleName(columnName));
 
         if (column == null) {
             return -1;
